Abstrayendo React Router, creando history y haciendo intialState mas accesible
